refactor(execution): extract buildResolveInfo helper
This is a BREAKING CHANGE for anyone using deep imports to import buildResolveInfo from execute.

Motivation: to break a cycle discovered by eslint.
